+++ R/add.R
@@ -338,8 +338,9 @@ addPlots <- function(study, plots, overwrite = FALSE) {
 #'   4) \code{labelLow} - The left-side label to describe low values of the statistic.
 #'   5) \code{labelHigh} - The right-side label to describe high values of the statistic.
 #'   To share metadata across multiple models, use the model ID "default".
+#'   6) \code{featureDisplay} - The feature variable to use to label the barcode plot
+#'   on hover.
 #' @inheritParams shared-add
-#'
 #'
 #' @export
 addBarcodes <- function(study, barcodes, overwrite = FALSE) {